Update 4: Add flash item and moving through 2 portals after broken.
 change color of player 2, fix bugs besides.

+++ src/uet/oop/bomberman/entities/Bomb/Bomb.java
@@ -150,7 +150,7 @@ public class Bomb extends Entity {
             this.img = Sprite.movingImage(ImageLists.bombImages.get(0),
                     ImageLists.bombImages.get(1),
                     ImageLists.bombImages.get(2),
-                    (int) BombermanGame.getTime(), TIME_FOREACH_IMAGE);
+                    BombermanGame.getTime(), TIME_FOREACH_IMAGE);
         }
 
     }
